# src/swagger.json
{
    "openapi":"3.0.0",
    "info": {
        "title":"API do Challenger UBI",
        "description":"",
        "termsOfService": "",
        "contact":{
            "email":"lucasamorimlima17@gmail.com"
        },
        "version":"1.0.0"
    },
    "servers": [
        {
            "url":"http://localhost:3333/",
            "description":"API de test"
        }
    ],

    "paths": {
        "/users/login":{
            "post":{
               "summary": "Autenticação do usuário",
               "description": "Rota responsável pela autenticação do usuário ultizando JWT",
               "tags": ["Users"],
               "requestBody":{
                   "content":{
                       "application/json":{
                           "schema":{
                            "$ref": "#/components/schemas/loginUsers"
                           },
                           "examples": {
                                "Users": {
                                    "value":{
                                        "email": "lucas@gmail.com",
                                        "password": "lucas"
                                    }
                                }
                            }
                        }
                    }
               
                },
                "responses":{
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type": "object",                
                                        "$ref":"#/components/schemas/loginUsers"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/add":{
            "post":{
               "summary": "Cadastra um usuário",
               "description": "Rota responsável por inserir um usuário",
               "tags": ["Users"],
               "requestBody":{
                   "content":{
                       "application/json":{
                           "schema":{
                            "$ref": "#/components/schemas/Users"
                           },
                           "examples": {
                                "User": {
                                    "value":{
                                        "email":"lucas@gmail.com",
                                        "password":"lucas",
                                        "role":"adm"
                                    }
                                }
                            }
                        }
                    }
               
                },
                
                "responses":{
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type": "object",                
                                        "$ref":"#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/task/add":{
            "post":{
                "summary": "Cadastra uma Task",
                "description": "Rota responsável por inserir uma task",
                "tags": ["Task"],
                "security" : [{"bearerAuth":{}}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                             "$ref": "#/components/schemas/Task"
                            },
                            "examples": {
                                 "Task": {
                                     "value":{
                                        "name":"task",
                                        "description":"description",
                                        "term":"2022-01-28T21:20:38.598Z",
                                        "finalized": false,
                                        "userId":"d0dcc155-77bc-4382-b21c-41ab29728797"
                                     }
                                 }
                             }
                         }
                     }
                
                 },
                 
                 "responses":{
                     "401": {
                         "description": "Unauthorized"
                     },
                     "200": {
                         "description": "OK",
                         "content": {
                             "application/json":{
                                 "schema":{
                                     "type": "object",                
                                         "$ref":"#/components/schemas/Task"
                                 }
                             }
                         }
                     }
                 }
            }
        },
        "/users/task/finalized/{id}":{
            "put":{
                "summary": "Finaliza uma task pelo id",
                "description": "Finaliza uma task pelo id",
                "tags": ["Task"],
                "security" : [{"bearerAuth":{}}],
                "parameters":[
                {
                     "name": "id",
                     "in": "params",
                     "description":"id da task para busca",
                     "required": true
                
                 }
                 ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                             "$ref": "#/components/schemas/finalizedTask"
                            },
                            "examples": {
                                "Update": {
                                    "value":{
                                        "dateFinalized": "2022-01-27T21:21:30.000Z"
                                    }
                                 
                                }
                            }
                        }  
                    }
                
                }
            }
        },
        "/users/task/update/{id}":{
            "put":{
                "summary": "Atualiza uma task pelo id",
                "description": "Atualiza uma task pelo id",
                "tags": ["Task"],
                "security" : [{"bearerAuth":{}}],
                "parameters":[
                {
                     "name": "id",
                     "in": "params",
                     "description":"id da task para busca",
                     "required": true
                
                 }
                 ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                             "$ref": "#/components/schemas/updateTask"
                            },
                            "examples": {
                                "Update": {
                                    "value":{
                                        "description":"description updated 9",
	                                    "term": "2022-01-28T21:20:38.598Z"
                                    }
                                 
                                }
                            }
                        }  
                    }
                
                }
            }
        },
        "/users/task/find/{id}":{
            "get":{
                "summary": "Dados das Places",
                "description": "Rota responsável por exibir os dados do place",
                "tags": ["Task"],
                "security" : [{"bearerAuth":{ }}],
                "parameters":[
                {
                     "name": "id",
                     "in": "params",
                     "description":"Nome do user para busca",
                     "required": false
                
                 }
                 ],
                 "responses":{
                     "401": {
                         "description": "Unauthorized"
                     },
                     "200": {
                         "description": "OK",
                         "content": {
                             "application/json":{
                                 "schema":{
                                    "type": "array",                
                                        "$ref":"#/components/schemas/Task"        
                                 }
                             }
                         }
                     }
                 }
            }
        },
        "/adm/find/":{
            "get":{
                "summary": "Dados de todos os usuários",
                "description": "Rota responsável por exibir os dados de todos os usuários",
                "tags": ["Administrator"],
                "security" : [{"bearerAuth":{}}],
                "parameters":[
                {
                    "name": "page",
                    "in": "query",
                    "description":"numero da página",
                    "required": false
                
                 },
                 {
                    "name": "perPage",
                    "in": "query",
                    "description":"numero de itens de uma página",
                    "required": false
                
                 },
                 {
                    "name": "search",
                    "in": "query",
                    "description":"filtro para busca busca",
                    "required": false
                 }
                 ],
                 "responses":{
                     "401": {
                         "description": "Unauthorized"
                     },
                     "200": {
                         "description": "OK",
                         "content": {
                             "application/json":{
                                 "schema":{
                                    "type": "array",                
                                        "$ref":"#/components/schemas/Task"        
                                 }
                             }
                         }
                     }
                 }
            }
        }
        
    },"components":{
        "schemas":{
            "loginUsers":{
                "type": "object",
                "properties": {               
                 "email":{
                     "type": "string" 
                 },
                 "password":{
                     "type": "string" 
                 }
                }
            },
            "Users":{
                "type": "object",
                "properties": {               
                 "email":{
                     "type": "string" 
                 },
                 "password":{
                     "type": "string" 
                 },
                 "role":{
                    "type": "string" 
                }
                }
            },
            "Task":{
                "type": "object",
                "properties": {
                    "name":{
                        "type": "string" 
                    },
                    "description":{
                        "type": "string" 
                    },
                    "term":{
                        "type": "string"
                    },
                    "finalized":{
                        "type": "boolean" 
                    },
                    "userId":{
                        "type": "string" 
                    }
                }
            },
            "updateTask":{
                "type": "object",
                "properties": {
                    "description":{
                        "type": "string" 
                    },
                    "term":{
                        "type": "string" 
                    }
                }
            },
            "finalizedTask":{
                "type": "object",
                "properties": {
                    "dateFinalized":{
                        "type": "string" 
                    }
                }
            }
        },
        "securitySchemes":{
            "bearerAuth":{
                "type": "http",
                "scheme": "bearer",
                "bearerFormat" : "JWT"
            }
        }
    }
}